dom/base/test/test_meta_viewport1.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>meta viewport test</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <meta name="viewport" content="width=device-width, initial-scale=1">
  <script src="viewport_helpers.js"></script>
</head>
<body>
  <p>width=device-width, initial-scale=1</p>
  <script type="application/javascript">
    "use strict";

    add_task(async function test1() {
      await SpecialPowers.pushPrefEnv(scaleRatio(1.0));

      let info = getViewportInfo(800, 480);
      is(info.defaultZoom, 1,    "initial zoom is 100%");
      is(info.width,       800,  "width is the same as the displayWidth");
      is(info.height,      480,  "height is the same as the displayHeight");
      is(info.autoSize,    true, "width=device-width enables autoSize");
      is(info.allowZoom,   true, "zooming is enabled by default");

      info = getViewportInfo(900, 600);
      is(info.width,       900,  "changing the displayWidth changes the width");
      is(info.height,      600,  "changing the displayHeight changes the height");
    });

    add_task(async function test2() {
      await SpecialPowers.pushPrefEnv(scaleRatio(1.5));

      let info = getViewportInfo(900, 600);
      is(info.defaultZoom, 1.5,  "initial zoom is 150%");
      is(info.width,       600,  "width equals displayWidth/1.5");
      is(info.height,      400,  "height equals displayHeight/1.5");
    });
  </script>
</body>
</html>
